+++ components/mip-date-picker/example/index.html
@@ -17,9 +17,15 @@
 
 <body>
   <p>date-pick 基本用法</p>
-  <!-- date-pick -->
   <mip-date-picker mode="date-picker"></mip-date-picker>
   <!-- <mip-date-picker mode="date-picker" min-date="2019-08-02" max-date="2019-08-26"></mip-date-picker> -->
+
+  <p>range-pick 基本用法</p>
+  <!-- <mip-date-picker mode="range-picker" on="select:test.select"></mip-date-picker> -->
+
+  <mip-date-picker id="test" mode="range-picker"></mip-date-picker>
+
+  aaaaa
 
   <!-- range-pick -->
   <!-- <mip-date-picker mode="range-picker" min-date="2019-08-02" max-date="2019-08-26">
